App/backend/requirements.txt
fastapi
uvicorn
scikit-learn
Pillow
matplotlib
numpy
scikit-image
joblib
python-multipart